+++ R/rerun.R
@@ -18,19 +18,17 @@
 #'   rerun(x = rnorm(5), y = rnorm(5)) %>%
 #'   map_dbl(~ cor(.x$x, .x$y))
 rerun <- function(.n, ...) {
-  dots <- substitute(list(...))
-
-  out <- vector("list", .n)
+  dots <- lazyeval::lazy_dots(...)
 
   # Special case: if single unnamed argument, insert directly into the output
   # rather than wrapping in a list.
-  if (nargs() == 2 && is.null(names(dots))) {
-    dots <- dots[[2]]
+  if (length(dots) == 1 && !has_names(dots)) {
+    dots <- dots[[1]]
   }
 
+  out <- vector("list", .n)
   for (i in seq_len(.n)) {
-    out[[i]] <- eval(dots)
+    out[[i]] <- lazyeval::lazy_eval(dots)
   }
-
   out
 }