+++ NEWS.md
@@ -25,3 +25,5 @@
 * `update_list()` can now modify an element called `x` (#98).
 
 * New infix attribute accessor `%@%` (#69).
+
+* `rerun()` is correctly scoped (#95)